devop: update packages, move component to core
