+++ package.json
@@ -38,10 +38,10 @@
     "@ensdomains/dnsregistrar": "0.4.0",
     "@myetherwallet/mew-components": "0.5.75-beta",
     "@myetherwallet/mewconnect-web-client": "2.1.23-beta.9",
-    "@sentry/browser": "6.4.1",
-    "@sentry/integrations": "6.4.1",
-    "@sentry/tracing": "6.4.1",
-    "@sentry/vue": "6.4.1",
+    "@sentry/browser": "6.5.1",
+    "@sentry/integrations": "6.5.1",
+    "@sentry/tracing": "6.5.1",
+    "@sentry/vue": "6.5.1",
     "@unstoppabledomains/resolution": "3.0.0",
     "@walletconnect/browser": "1.0.0",
     "@walletconnect/qrcode-modal": "1.4.1",
@@ -56,7 +56,7 @@
     "apollo-utilities": "1.3.4",
     "bitbox02-api": "0.8.1",
     "clipboard-copy": "4.0.1",
-    "core-js": "3.12.1",
+    "core-js": "3.13.1",
     "graphql": "15.5.0",
     "graphql-anywhere": "4.2.7",
     "graphql-tag": "2.12.4",
@@ -68,8 +68,8 @@
     "platform": "1.3.6",
     "pocket-js-web3-provider": "0.0.3",
     "register-service-worker": "1.7.2",
-    "subscriptions-transport-ws": "0.9.18",
-    "vue": "2.6.12",
+    "subscriptions-transport-ws": "0.9.19",
+    "vue": "2.6.14",
     "vue-apollo": "3.0.7",
     "vue-router": "3.5.1",
     "vuex": "3.6.2",
@@ -144,9 +144,9 @@
     "u2f-api": "1.2.1",
     "uglify-es": "3.3.9",
     "uuid": "8.3.2",
-    "vue-cli-plugin-vuetify": "2.4.0",
+    "vue-cli-plugin-vuetify": "2.4.1",
     "vue-i18n": "8.24.4",
-    "vue-template-compiler": "2.6.12",
+    "vue-template-compiler": "2.6.14",
     "vuetify": "2.4.3",
     "wallet-address-validator": "0.2.4",
     "web3": "1.3.6",
